highway_api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
PyMySQL==1.1.0
openai>=2.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
